refactor: [ANDROAPP-7138] Adapt integration test to algin with Gherkins definition
